break apart AccountSettings to its own directory with files for each feature
